Rename and move the classes
